Revert "fix(oauth): cross-origin PKCE — opener exchanges code, not popup (#46)"
This reverts commit 8fdbca31d6193da98efb9499b24819be43d67dd3.
